Lo logré

+++ AppCiudades/app/src/main/java/com/example/appciudades/dominio/models/Cerveza.kt
@@ -2,6 +2,7 @@ package com.example.appciudades.dominio.models
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.google.android.gms.maps.model.LatLng
 
 
 @Entity
@@ -10,7 +11,7 @@ data class Cerveza(
     val nombre: String,
     val ciudad: String,
     val pais: String,
-    val grados: Int,
+    val grados: Double,
     val latitud: Double,
     val longitud: Double,
 //    val img: Image?,
